Add base files for first commit
The basic server supports multiple client connections by using the
select() command that manages all created client connections.

It loops through each connection created, and checks if each is ready to
read from (or write to). If ready, then normally means that client is
sending information to the server. If the file descriptor ready to be
read is the server socket itself, then that means that a new client is
ready to connect.