Attempt to fix HTTP header problem - not fixed
The server reads sockets only when they have data ready to send, but it
seems nc and telnet send their http request data line by.

Currently the program takes in the "GET / HTTP/1.1" line and then sends
it to the server, clearly getting an unknown response.

Perhaps need to only read server response when it is ready to send
something back. It may only reply unknown because we are ready from it -
if we send all the data incrementally first then it may put it together.