Get Telnet and NC working fully
Now noth Telnet and NC work with HTTP requests, with a double \r\n to
get the HTTP response. They can be both disconnected at the remote host
and at the client and the server does not crash.

To do: Ensure that all vectors passed as parameters are passed by
reference, rather than duplicated, when necessary.