Finish basic prototype of find replace function
A basic working prototype of the find replace function is completed.

It replaces a certain string with a longer one in the message, by
dynamically increasing the size of the main message string as necessary.

To do so, realloc is necessary. Normally when changing a string in a
function, the values pointed to by the char * pointer are changed.
However, with a realloc the actual memory location is changed, not just
the values in it.

Therefore, it is necessary to pass the pointer to the char *, rather
than the char * itself, and update this directly when a realloc occurs.

To do: Clean up find_replace function and add support for replace large
string with smaller replacement. Also parse -replace command line
options instead of hard coding them.