Add find replace functionality to message receive
Now the specified find replace functionality has been added to the
receive_message function.

So if div --> YEAH! then HTTP response is changed to <YEAH! ... </YEAH!>
etc.

To achieve this, it was necessary to pass a char ** to the
receive_message function to allow for pointer reallocation in the find
replace function.